server/requirements.txt
pandas
numpy
faiss-cpu
openai
python-dotenv
statsmodels